getFullState(), deposited/burned/release tracker vars, and comments
Added in the three units fundsDeposited, fundsBurned, fundsReleased. Pretty self-explanatory; it tracks how the ether has flowed through the BOP. Displaying these, along with the BOP's remaining balance, gives the user a very good feel for the "progress" or "completion" of a given BOP.

Added getFullState(). It is a bit of a misnomer, as it really only returns the parts of the state that are able to change. So it doesn't return payer or commitThreshold, for example, because a given BOP will never change either of these variables. This function also returns the above three members. Calling getFullState() rather than fetching each variable getter independently can be way faster on systems which experience any kind of lag in communication between an application and the Ethereum node.

The above two changes relieves the client of having to independently track/calculate how much has been burned, released, or deposited to the BOP.

Also added a lot of comments near the beginning to describe the BOP variables.